rewrite of Parse and Validate (flags only, not args)
 
this branch:
$ ts-node test/perf/parser.perf.ts
simple x 131,268 ops/sec ±1.15% (79 runs sampled)
**multiple async flags that take time x 9.90 ops/sec ±0.14% (50 runs sampled)**
flagstravaganza x 7,425 ops/sec ±2.10% (83 runs sampled)

main
$ ts-node test/perf/parser.perf.ts
simple x 103,392 ops/sec ±0.91% (77 runs sampled)
**multiple async flags that take time x 4.96 ops/sec ±0.13% (28 runs sampled)**
multiple async flags that take time x 5,451 ops/sec ±2.09% (79 runs sampled)

lots of refactoring was needed to enable parallelization, but the main goal is to run flags' async parse methods in parallel.

The code could be a lot cleaner if we could make the assumption that this.context on Parser is not mutating.

That middle perf test `multiple async flags that take time` is meant to model something that's got some delay on it (ex, various local fs operations like sf's Org flag)