@WillieRuemmele 2023-07-11
### QA Notes
---
✅ : validated with NUTs
✅ : in PRM with this built and linked required flag
```bash
 ➜  ../../oss/plugin-release-management/bin/dev github:check:closed                                                   
Error (1): The following error occurred:
  Missing required flag github-token
```

✅ : read flag value 
```bash
 ➜  ../../oss/plugin-release-management/bin/dev github:check:closed --github-token abc
Error (1): Bad credentials
```

✅ : required flag validated 
```bash
 ➜  ../../oss/plugin-release-management/bin/dev github:check:closed                                                   
Error (1): The following error occurred:
  Missing required flag github-token
```

✅ : read from env 
```bash
 ➜  GITHUB_TOKEN=abc ../../oss/plugin-release-management/bin/dev github:check:closed
Error (1): Bad credentials
```

✅ : relationship flags still validated
```bash
 ➜  ../../oss/plugin-deploy-retrieve/bin/dev project:deploy:start -d force-app --results-dir abc                      
Error (1): The following error occurred:
  One of the following must be provided when using --results-dir: --coverage-formatters, --junit
```